Broke down JS file into subfiles
