+++ index.html
@@ -56,6 +56,7 @@
       href="https://fonts.googleapis.com/css2?family=Fira+Code:wght@500&display=swap"
     />
     <script type="module" src="./script.js" defer></script>
+    <script type="module" src="./scripts/eventListeners.js" defer></script>
     <title>Terminal</title>
     <link rel="icon" type="image/jpg" href="favicon.ico" />
     <link rel="stylesheet" id="switcher-id" href="" />
